application: syncfs on storage during startup
If Redpanda crashes during storage writes, there
may be non-persistent metadata left visible in
the filesystem.

To provide the guarantee that everything the storage
layer sees on disk during recovery is persistent,
call syncfs during startup.

The strength of this guarantee depends on the underlying
filesystem, as some filesystems may lie about persistence
if there has previously been an I/O error on a device,
by marking pages clean even though they were not
successfully flushed.

This does not help in all cases, but it is a cheap
thing to do that may help us reason about storage
consistency issues across crash/restart cycles if
they are seen in the field.