rptest: support finjections in RedpandaService
This commit extends RedpandaService with a member function that deploys
a failure injection configuration to Redpanda nodes. When this is used,
RedpandaService will not assert out on node crashes. Failure injection
can result in assertions being triggered on EIOs, so that needs to be
tolerated. Test authors should use this functionality in conjunction
with the RedpandaMonitor service that restarts crashes nodes.