Use explicit ref type for transform iterator
In fetch_session we used a boost::transform_iterator to create an
iterator over fetch partitions on top of a collection of fetch
metadata stored in the fetch session (i.e., a "projection" down to
the partition).

This happened using a lambda with an implicit auto return type, like:

[](const entry& e) {
    return e.partition;
}

However, auto will deduct non-reference types by default, so this was
equivalent to:

[](const entry& e) -> kafka::fetch_session_partition {
    return e.partition;
}

This means copying the partition object on every access of the iterator,
which involved an allocation.

Instead, just return a reference to the contained partition object,
by setting the return type explicitly:

[](const entry& e) -> const kafka::fetch_session_partition& {
    return e.partition;
}

Issue redpanda-data/core-internal#435.