Benchmark for fetch planning
Add an rp_bench for fetch planning.

This benchmark uses the redpanda fixture as the stuff it's benchmarking
needs a pretty fully fleshed out redpanda setup, but this seems to
work fine as long as we only enable timing around the loop that matters.

This tests the scenario of creating the fetch plan against a fetch
session, which has been shown to use a lot of CPU based on profiles
from customers and test environments.

Issue redpanda-data/core-internal#435.