storage: rename log::housekeeping to apply_segment_ms
Because we are splitting the compact() function into separate functions
for compaction() and gc(), we need a name that can be used to represent
the most common function which is the combination of the two without
updating every call site to call both primitives instead.

Housekeeping is the prefect name. So we renamed the previous
housekeeping method to reflect its exact functionality. Notice that in
the comments it is clear that in the future apply_segment_ms will likely
be folded directly into the housekeeping functionality itself, but this
commit does not attempt to address those TODO items.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>