cloud_storage: Add chunk api and data source
A chunk API and data source are introduced. The chunk API is per remote
segment. It is used to manage a map of chunk start offsets to metadata.
It also manages trimming of unused chunks and retrying chunk hydration.

The chunk hydration is managed via remote segment, the chunk API retries
continuously in case of errors during hydration. The chunk API also
provides an interface to find the next chunk as well as to iterate over
chunks in segment.

The chunk data source accepts a range of chunks and reads through the
range, hydrating the chunks as necessary by calling into the chunk API.
It is responsible for translating the kafka offsets in a fetch request
to chunks using helpers in remote segment.

The remote segment uses a coarse index which is an index with low
resolution mapping kafka offsets to file offsets to assist the chunk
api in initializing the metadata for each chunk. It also gets new
method to hydrate a chunk on demand.

The chunk API is a member of the remote segment and is started lazily
after the index is downloaded, this is because the coarse index cannot
be generated until the index is downloaded, and the coarse index is
needed to create the chunk metadata map.

The remote segment is also extended to contain the concept of a fallback
mode. In this mode, either the cloud segment version is old enough not
to have an index, or the index download has failed. In either case the
segment is fully downloaded and chunks are not downloaded.

Once fallback mode is engaged, it is set for the lifetime of the remote
index.