cloud_storage: Decrease write buffer size in c-store
The max_buffer_entries was set to 1024 which is too high. This value
makes size computation for the c-store unpredictable because a lot of
elements are stored in the write buffer without being written to the
columns. The value is set to 64 which is a size of one column frame.