Use only mq mode for NIC tuner
Currently we try to choose among sq[-split] and mq modes for assigning
NIC IRQ affinity. However the idea behind sq and sq-split modes is that
a core is *dedicated* to IRQs (i.e., Redpanda core does not run on that
core or lcore). Nnow, we don't currently support propagating a cpuset to
Redpanda, so RP will run on the IRQ core, causing it to be the first
bottleneck, at least for high packet count loads which stress softirqs.

We change the tuner to use mq mode exclusively for now, regardless of
the number of cores or lcores. In the future we may investigate the use
of the sq modes again, if we can propagate a cpuset to Redpanda.

Fixes issue #10444.