storage: use a long-lived fiber for housekeeping
Prior to this commit storage housekeeping was run as a one-shot fiber
started from a periodic timer. That process may have taken quite some
time given that it did a round of compaction and retention GC on all
partitions.

In low-disk situations it is desirable to respond asap by reclaiming
space. However, the one-shot fiber may have be doing non-impactful work
when the situation arises. We need a mechanism to reschedule work with
low latency.

One solution is to start another background fiber on demand for
important work, but this adds unncessary complexity when it comes to
making sure that the work of each fiber is not interfering with the
other (e.g. both may be performing similar operations).

Instead, this commit transforms the housekeeping loop into a long-lived
fiber in which a single policy can be implemented, and avoid the
complexity arising with a solution based on concurrent worker fibers.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>