topic_recovery: make status check when starting atomic
CONFLICT:
- kept std::move(request) as exists in this branch

We currently return immediately if a topic recovery isn't active on the
current node, without setting the recovery state to indicate that a
recovery is now running. This currently happens in a background fiber,
which makes this prone to races.

This commit moves the state change out of the background fiber so the
state is changed before returning to clients.

(cherry picked from commit eccb6f4abef87183792d3021fb7b3aadd7a54963)