topic_recovery: make status check when starting atomic
We currently return immediately if a topic recovery isn't active on the
current node, without setting the recovery state to indicate that a
recovery is now running. This currently happens in a background fiber,
which makes this prone to races.

This commit moves the state change out of the background fiber so the
state is changed before returning to clients.