partition_manifest: implemented serde serialization
in the ::update() path, json and serde format are accepted. on the
::serialize() path, only serde is used. The caller is responsible to
make sure that the upload of a serialized manifest is done behind a
feature gate

get_legacy_manifest_path returns manifest.json
it's intended to be used to retrieve the old format

whenever a partition_manifest is downloaded, first it' tried the binary
format then the json format.

in contrast, only binary format is used for uploading, but the action
itself is behind a feature flag check.

json serialization is retained for admin api and debug output