archival: report dynamic metrics from the leader
This commit reworks the archival probe to distinguish between static and
dynamic metrics. Dynamic metrics represent the current state of the
archiver and should only be reported by the leader. Static metrics
provide a historical view of the work that has been performed and are
reported by all members; these metrics should be summed-up by node.

The a ref to the `ntp_archiver` has also been added to the probe. This
will be used by future commits.