k/fetch: soft assert on diff sizes of `shard_fetch` collections
When the `shard_fetch` collections (results, responses, and metrics)
get to `fill_fetch_responses`, soft assert the invariant that
they have to be the same size. If they are not, only process
the common subset of those. That may result in empty fetch responses,
which is bad, but probably better than segfaulting.