cloud_storage: Handle overlapping segments
When a reader appears stuck because it cannot make progress through its
current segment, one reason for this may be that the segment ends
earlier than the manifest entry indicates, and the next offset to be
read from lies within the next segment. This change catches a stuck
reader exception, and if the reader is not at the logical end but the
next offset lies in the next segment, the reader is forced to an EOF
state. Then we try to reset the reader to point to the next segment,
which should allow the read to progress.