storage: report minimum capacity needed for logs
When collecting a storage report, we add a new section that descibes
target sizes with different meanings. This commit adds a minimum
capacity target which is the minimum amount of capacity that log storage
needs to be able function at a bare minimum.

The minimum is computed as the max segment size (potentially different
for each partition) * the minimum reserved number of segments
(configurable) * the total number of partitions.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>