cloud_storage: Adjust conditions to reset stuck reader
When a reader is stuck before reaching EOF and the end offset of the
segment overlaps the next segment, we reset the reader. This condition
is fixed to include the overlap. The change made recently was not enough
to cover overlapping segments correctly.

	modified:   src/v/cloud_storage/remote_partition.cc

(cherry picked from commit 600f6ee8def3532aa7d2f99daaf57475a6535451)