k/quotas: use "kafka_throughput_control" for tput exemptions
connection_context now has all quota related stuff for the connection
stored in `_snc_quota_context`. This object is supposed to be created
once per connection lifetime by `snc_quota_manager`, but it will be
recreated each time a client_id changes on the connection.

When the quota context is created (lazily on the connection context),
the `kafka_throughput_control` rules are used to select the matching
throughput control group. If any group is matched, the context saves it
as a flag to exempt the connection from any snc_quota_manager control.
This will change into a full association with the control group.
Currently the exempt flag simply tells the quota manager to skip any
messages in that context.