rptest: harden final storage consistency check
A consistency check runs against each Redpanda node at the end of
ducktape tests. It queries the `local_storage_usage` admin endpoint and
cross checks it against the size of the data directory.

Previously, this check did not take into account the fact that Redpanda
can apply retention between the local and remote query. This commit
updates the check to fudge the maximum percentage diff by the amount of
data reclaimable by retention.