rptest: test retention migration mid-way through
When a cluster is upgraded to v22.3, the semantics of retention.ms|bytes
change. retention.local.target.ms|bytes now controls retention on the
local log, so the previous values are migrated into the new configs.
This should be the case even during the upgrade process (i.e. not all
nodes have been upgraded to v22.3 yet).

This commit introduces a test to validate this behaviour.