cloud_storage: fix empty segment check in reader
This was incorrectly offsetting the next kafka offset
by 1.

A segment with kafka data in it has a next offset != its
base offset, while a segment _without_ kafka data has a next
offset == its base offset.

This was ignored by unit tests because the tests were
using segment name format v1 which ignores delta_offset_end,
and ignored by integration tests because if our reader path
drops out then the client will retry, and because it only
causes the problem if a segment has exactly one data
batch in it.