r/consensus: prefer availability over voter priority
Redpanda Raft implementation exposes an API allowing to override a voter
priority. This is used by the drain manager when a node is in
maintenance mode. In current implementation when the only voter is in
maintenance mode the Raft group is not able to elect a leader as the
reported priority it to low (the priority override in maintenance is set to 0).

Fixed Raft implementation to make sure that it prioritize an
availability over the user priority preference. If a node is the only
voter the priority override is ignored.

Fixes: https://github.com/redpanda-data/cloudv2/issues/6174

Signed-off-by: Michal Maslanka <michal@redpanda.com>