Change move strategy for ktp_with_hash
Initially, ktp_with_hash would recalculate its cached hash code when
it was the rhs for a move ctor or assignment: since this mutates the
object (specifically, the topic name becomes empty as it is a string)
the cached hash code is invalid.

As pointed out in review, this is perhaps not worth the cost in
complexity and performance. Instead, we leave the hash code as-is,
using defaulted move operations and document that you should not call
the hash method in a moved-from state.

We enforce this rule in non-release builds with an assert using the
moved-from canary. We also check that the base ktp object is not in
a moved-from state when either the topic or partition accessors are
called.