rm_stm: optimize lso for acks=1 case
In general case LSO can't go beyond applied offset but it reduces
acks=1 e2e scenarios because a record only becomes visible when its
offset is behind LSO but LSO depends on applied offset which depends
on a record being fully replicated so reading acks=1 is as slow as
reading acks=all.

A leader knows about the all inflight replication requests so it
may return LSO beyond applied offset when it knows that it can't
be affected by the inflight requests.