rpk: remove tuners from rpk.yaml
It's difficult to reason about whether to error or not if one of the
redpanda.yaml or rpk.yaml fails to be written. Keeping tuners strictly
in redpanda.yaml isn't too bad of a papercut, and this allows us to
wait some time to figure out what we want to do in the future.

In preparation of the future, though, SetMode has been moved to the
Config struct. As well, we now autocomplete the two modes supported.