r/consensus: use separate mutex to protect snapshot
When writing a snapshot Raft uses the state that is a part of already
committed log, offset translator and configuration manager. The snapshot
does not relay on the protocol state that mutates. This observation
allows to write snapshot outside of Raft operation mutex as that
operation is not changing any of the current state and does not relay on
the current state. Using separate mutex to protect the snapshot reduces
contention and does not influence the write path latency.

Signed-off-by: Michal Maslanka <michal@redpanda.com>