add pre-exit hook that calls the vtools one
Currently the bk hooks are defined in the vtools
repo and they are moved to the redpanda pipeline
at runtime. We shouldn't move hooks in CI as it
causes git to report a diff. This causes the
-dirty suffix to be present when running
redpanda --version

ref
https://github.com/redpanda-data/vtools/issues/1753
https://github.com/redpanda-data/vtools/pull/1760#discussion_r1198914948

(cherry picked from commit 09759459800a0c1850415dcd33e2a49f37b82f05)