tm_stm: prevent double hosted txs initialization
try_init_hosted_transactions check if hosted transactions are initialized
and if they don't try_init_hosted_transactions takes the write lock and
initilized the transactions.

However while the method waits for the write lock the state may change so
to avoid hosted txs from being initialized twice it's necessary to check
the state again after the lock is taken.