storage: speed up cloud storage status calculation
Previously we calculated the overlap between local and
cloud storage, which required iterating over almost all
local segments (those which are also present in cloud
storage).

Instead, just iterate over the most recent segments
which are _not_ in cloud storage: this is typically
just 1-2 per log.

On a healthy system this makes the calculation O(1)
instead of O(N), although on systems that are far
behind with uploads this will still be O(N).