cloud_storage: Handle nested benign exception
If a nested exception is thrown where either the inner or outer
exception is abort requested or gate closed, this is logged as a debug
level error.

If the nested exception is of any other kind, an error level log is
generated.

(cherry picked from commit 69b4c6918e8f0bd395576d5ecfd8b635b82318b9)