kafka: include remote storage in DescribeLogDirs
Tiered storage topics are reported in two directories: the
local storage path, and a remote storage path.

The directory for remote storage is given as "remote://{bucket}"

The two reported sizes do not add up to total space used: they overlap.
The reason we report both separately rather than an overlap-resolved
total is to serve the operational purpose of answering "how full is my..."
questions.  If we reported a total usage, then it wouldn't help
a user who wanted to keep track of which topics were eating their
local disk space.

Fixes: https://github.com/redpanda-data/redpanda/issues/8145