tests/utils: add quiesce_uploads, fix revision loading
quiesce_uploads: this is a helper for any function that wants to wait
for all local data to be uploaded to S3 before proceeding.

revisions: we were previously incorrectly using admin API's _current_
revision for a partition, we should have been using the initial
revision.  We can get this from S3 without touching the live cluster.