cloud_storage: track hydration loop status
This commit extends `remote_segment` to track the status of the
hydration loop. Previously, if a hydration is requested via
`remote_segment::hydrate()` whilst the loop is not running, the request
will stall until the client times it out. This also prevents Redpanda
from cleaing up the Kafka connection properly.

With this patch, the client gets an error instead an the connection is
released.