cloud_storage: Add chunk range for prefetching
A chunk range covers a series of contiguous chunks, which can be used to
prefetch. The idea when prefetching is that we need to make a single http
call for a byte range which contains > 1 chunks, but we want to write the
data to disk as separate chunk files. The range utility introduced here
takes a start offset, a prefetch count and a map of chunk starts, and
enables iteration over the chunks until the prefetch is satisfied, also
allowing access to the last offset which is used to download the byte
range.