cloud_storage: Add bounded stream
A bounded stream wraps an input stream and only consumes upto a certain
offset from the underlying input stream. It is intended to be used with
chunk prefetch, where a single input stream is used to read the response
from an http GET call, and bounded streams are created on it to read up
to chunk boundaries to be written to disk.