cloud_storage: Add stream consumer for prefetch
The remote object expects a consumer which will take an input stream and
do something with it (usually write the response to disk). The consumer
is supposed to be re-entrant since it can be called multiple times by
remote. This change adds a stream consumer which accepts an http
response stream and a chunk range, and creates chunk files to put in
cache from the stream containing potentially multiple contiguous
chunks. It effectively breaks up the stream into chunks before putting
the files in cache.