archival: simplify handling of segment merging enable/disable
It was somewhat fragile to rely on an external caller
to always use set_enabled() to toggle segment merging in
a way that takes account of the cluster config _and_ the
state of the system.  For example the previous commit fixed
a bug where segment merging could get turned on even if it
was configured to be off.

Instead, let's separate the set_enabled() path and the use
of cluster config, have the merger own its own config binding
for the latter, and use && at runtime to decide whether to
act.

Logically, the two configs are:
- Cluster config: when I run, should I do anything?
- Job config: should I run at all?