features: add `new_clusters_only` mode
This is like features that auto-activate, but they
only auto-activate on clusters created with the
version >= the feature's required version.

This is suitable for features that we have good confidence
in, but don't want to impose on existing stable clusters
yet.

This policy can be changed in-place to `always` in later
.z releases at will, when we would like to activate
the feature on all clusters, irrespective of original
version.