cloud_storage: Add separate code path for spillover
Previously, the manifest truncation and spillover were performed using
the same `truncate_cmd` command in the archival STM. This caused space
accounting problem because both normal truncation and spillover
increased archive size paramterer. This commit adds new command similar
to truncate_cmd and separate code path that handles it in the STM and
the manifest. The spillover code path increases the archive size and the
normal truncation doesn't increase it. This solves the issue of space
accounting.