c/planner: introduced priority tiers for moving partitions on full disk
Introduced priority tiers for moving partitions when node disk is
getting full. The default policy is to move the smallest partition first
as they are the fastest to release the disk space from the node that is
at capacity. However when partition are smaller than than a certain
threshold it may not make much sense to move them first as the move
coordination would take longer than copying data between nodes.
Additionally the internal partitions like consumer offsets and tx
coordinator, are usually small and were moved with the high priority
introducing unnecessary disturbances in cluster operation.

Priority tiers allows organizing order in which the partitions are moved
when node disk is getting full. The highest priority is assigned to the
partitions with smallest size that is greater than threshold then in the
next step internal partitions are moved and finally the small partitions
that's size fall bellow the threshold.

Signed-off-by: Michal Maslanka <michal@redpanda.com>