config: add 'aliases' to property metadata
This enables a property to be known by more than
one name. This is necessary for backward compat when
we would like to evolve the name of a property
over time, whether it's to fix a typo, or to
reorganize and clean up naming more generally.

In config_store, aliases are used for singular get/contains
operations but not for listing operations.  That means
that at higher levels, we will permit setting properties
via their alias, but when listing out the properties
and their values, we will only include each property
once and via its primary name.