storage: allow waiting for cluster UUID
The cluster metadata uploader will need a cluster UUID to uniquely
identify the metadata it writes. To that end, it needs to be able to
wait for the cluster UUID to be generated.

This adds a method to storage::api to allow waiting for a call to
set_cluster_uuid().