rptest: tighten cloud deleted segments check
This test transfers leadership around, which can result in the partition
getting removed from a specific shard. This also includes removal of the
archiver and its probes. The assertion for the number of deleted
segments will fail if:
* partition move changes replica set completely
* no further deletions happen after the move

The issue is fixed in this commit by checking inferring deletion from
the manifest instead of the metric.