c/partition_balancer: split out full batch condition for cancellations
It can happen that we can't add any more reassignments but can add
cancellations, moreover it is necessary to add cancellations to make
progress (e.g. all 50 partition movements are already underway but are
targeting dead node so are unlikely to finish). Therefore we must
distinguish full batch conditions for reassignments and cancellations
and don't exit early if there are cancellations to be made.