tx registry: add empty tx_registry_stm
Intoducing tx_registry_stm a tx registry service will be based on.
tx_registry_stm is modeled after is_allocator_stm and doesn't rely
on cleanup policy to control the size of the log. Instead the stm
does prefix truncation on its own then the size of the log passes
_log_capacity batches.