cloud_metadata: start metadata ID counter at the highest
If starting an upload after the demise of a cluster, we need to be able
to differentiate between metadata from an older cluster and a newer
cluster. A recovery process must be able to pick the metadata from the
newer cluster upon restoring.

To that end, the metadata ID used by clusters restored in the same
bucket need to have monotonically increasing metadata IDs.

This commit alleviates this by tweaking the behavior when a cluster
uploads a manifest for the first time: it will now look through the
bucket for manifests from other clusters, and start their metadata ID
count from the highest seen.