tests/services: make metrics_sample more robust
Previously, if you passed in your started nodes,
then that list could be empty, metrics_sample
would replace it with self.nodes, and then
metrics() would assert out because it was trying
to run on non-started nodes.

Instead, if the incoming list is empty, tolerate
that, and just return an empty result.