rpk: change rpk's sasl "type" field to "mechanism", update flags
The proper terminology for SASL per the original RFC4422 is "mechanism".
Type is shorter but technically not correct, and our flags have always
referred to this as "--sasl-mechanism". We switch the new -X flag to be
"-X sasl.mechanism" to closely match the old flag, and we now encode the
SASL field with "mechanism". We still support decoding "type" but we
will rewrite it as "mechanism" when we encode.

This does also affect redpanda.yaml, meaning if a person modifies the
rpk section in redpanda.yaml, then older rpk's will no longer be able to
read their yaml. We historically do not support forward compatibility
like that (requiring old versions to work on new), and rpk will be
modifying rpk.yaml and not redpanda.yaml without some deliberate effort,
so this is perceived as low risk.