rpk config: add test to ensure rpk.yaml version is bumped properly
This at least ensures that if a person adds a field anywhere in
rpk.yaml, we force a version bump (in review). This also prevents
problems that we ran into with redpanda.yaml in the past, where a field
would be stripped by rpk because rpk's struct wasn't updated or an old
rpk was running. In redpanda.yaml, this was "fixed" by adding `Other`
basically everywhere. In rpk.yaml, we can do strict version checking:
new rpk's will support old rpk.yaml versions, old rpk's will fail
because of the version check up front, and this test forces us to bump
the version whenever appropriate.