rpc/transport: fix temporary dispatch loop stalls
Currently the way the code is structured can result in temporary
dispatch stalls timing out a bunch of requests.

Consider a request queue of sequence numbers [5, 6, 7, 8]

Lets assume seq=6 dispatch got delayed. This can happen if timeout is
low or there is an unknown delay by the scheduler (eg: debug builds).

_last_seq=4, queue = [5, 7, 8] - 5 is dispatched right away
_last_seq=5, queue = [7,8] -- out of order.

Dispatch of 7, 8 doesn't happen right away due to out of order sequence
and we never do a dispatch with seq=6 because we detect it already
timed out.

Now it takes a new RPC request to clear the stalled queue but
they may not happen for a while as most of these are timer based and
meanwhile seq=7/8 timeout for no fault of theirs.

This patch does two main things.

* Consolidates _last_seq tracking. Currently it is not monotonic and can
jump all over the place. This is confusing. Now we only update it in a
centralized place in dispatch_send().

* ^^ Requires that dispatch_send() is called in all cases which also
avoids dispatch loop stalls. For example, a timed out request will clear
the stalled queue right away (which is not the case before).

(cherry picked from commit 453d340c93357b234e4c736a1975033b8ae6613c)