tests: check for perfect replicas distribution
In the criteria to determine that the partitions are now balanced,
check whether max(replicas per node) is greater than min(replicas
per node) by no more than 1, in this case the partitions are considered
balanced without checking the 20% tolerance range.

This enables distributions like [1,1,2] to be always successful regardless
of the tolerance range.

Fixes #11042