c/partition_allocator: preserve shards for replicas on original nodes
After one or several reallocations a replica may end up on one of the
original nodes. Previously, we chose the shard for this replica anew in
this case. This is awkward because the shard could change so the full
replica id (i.e. broker_shard) changed even if the reallocation didn't
change anything in the grand scheme of things. So this commit makes
reallocations preserve the shard of replicas that end up on original
nodes.

And of course this change makes it easier in the future to let nodes
choose shards for the replicas they host themselves.