txn: step down on begin_tx & store_txn_offsets failures
Transactions in kafka protocol are stateful: the processing of the
requests depends on the previous commands executed by the same or
even different producer. It makes the situations when the replica-
tion fails with the indecisive errors such as timeout dangerous
because the true state is unknown.

Stepping down to resolve uncertainty by replaying the log

(cherry picked from commit 7ca57077f2a4c3ed592d4a989c11729f1aced332)