cloud_storage: Handle expected hydrate errors
When chunk hydration fails due to abort request, this change makes sure
that exception handling paths are mutually exclusive. There was a bug in
the previous code where the abort requested error was handled and then
control was handed off to the generic exception handler, resulting in an
unexpected error message.