utils: bind retry strategy to retry chain node
This commit binds the retry strategy to a given instance of a retry
chain node. The default behaviour is for, non-root nodes to inherit the
retry strategy from their parent. New constructors were added to permite
overriding of the retry strategy. Nothing should change for pre-existing
code that uses retry chain node.