cluster: drop redundant RPCs to drop lifecycle marker
Redundant RPCs might happen as a result of multiple
independent scrubbers processing the same topic, or
as a result of a single scrubber retrying.

It is always safe to ignore these: in the event that we
incorrectly left one behind, it would get hit by
some future iteration of az scrubber.

Fixes https://github.com/redpanda-data/redpanda/issues/11090