c/allocation_state: maintain final partition counts
When moving partitions, partition allocations on old nodes are not
removed until the move is finished. This makes sense because
partitions remain physically present on old nodes until the very end.
But we also use these allocation counts for count-based rebalancing.
This is incorrect because after all moves are finished, the distribution
won't necessary be uniform. To fix this, introduce "final counts" - the
number of partitions on a node after all currently in-progress moves are
finished and maintain it when (re)allocating partitions and in
topic_updates_dispatcher.