treewide: Mark all probes as immovable
Probes register metrics by capturing `this`, and it's not save to move
them after that. In a lot of places this is safe because their lifetime
is directly tied to a service which lives the whole program's lifetime,
but any small move of that object even during initialization can break
things (see #11155, #11095, #11107).

This takes a big hammer approach to removing this foot gun by making all
probes immovable and wrapping them in `std::unique_ptr`.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>