ducktape: Reduce parallel reads in test
The test changed in commit asserts that the read path works when the
segment size is smaller than chunk size. The expectations in test are
often unmet on debug build in CI, we expect to be able to read 2500
messages total but the test falls marginally short. Since this test is
not a load test but simply a check that read path works in this corner
case, this change lowers the message count for the test to pass.