cluster: remove inline cloud purge on ntp deletion
Previously, when an ntp was deleted, one of the replicas would do an
attempt to delete the segments pointed to by the current manifest and
the manifest itself.

Post spillover manifests this makes less sense as this best effort pass
does not attempt to delete all the data. Changing it to perform the full
purge could destabilise the cluster if deletetion occurs during a busy
period.

Instead, this commit removes the code path and deletion of cloud data is
performed only via the scrubbing mechanism, as that will only(1) kick in
when there's no outstanding requests to cloud storage.

(1) - if cloud_storage_housekeeping_interval_ms elapses without a scrub,
then it will run regardless of the system's state.