c/node_status_backend: add node to status table right after it joins
When node joins the cluster it should be added to the
`node_status_table` so that all the downstream systems relaying on the
node_status are aware that the node appeared in the cluster.

Previously adding node to status table was deferred until the first
successful heartbeat reply was received.

Signed-off-by: Michal Maslanka <michal@redpanda.com>