storage: remove last_term_start_offset from offset_stats
This value is only ever used in persisted_stm on term changes. And the
storage::log::offsets function that calculates it often shows up on perf
top. This PR reduces the time complexity of the function from being linear
to the number of segments since term start to being constant in most cases.