cloud_storage: write remote lifecycle markers after purge
The purging state is useful because is disambiguates "damage" like
missing segments from topics that are just being deleted right now:
any external reader like a bucket scanning tool or a read replica
will be interested in this distinction.

The purged state is useful because it provides a durable way for
external bucket scanning tools to confidently clean up any
orphan objects whose NT-revision is equal to that of a purged
lifecycle marker.