ducktape: fix installing confluent-kafka on arm64
On arm64 machines, there are no python wheels for confluent-kafka. When
attempting to install them with the previous Dockerfile, the librdkafka
source was not available. This commit uses the librdkafka intermediate
image as the base of the final stage of the ducktape test node image

Signed-off-by: Ivo Jimenez <ivo@redpanda.com>
(cherry picked from commit 4337602e630e614ff1f71ffee60c9c2c4a272235)