archival: fix idle timer scheduling
Previously, the idle timer would not fire if the shard had no
`cloud_storage::remote` activity. This was problematic because it
prevents topics from being purged when cloud storage is idling.

This commit fixes the issue by kicking the timer off in
`upload_housekeeping_service::start` and by re-arming in the idle timer
callback. Delaying the timer on cloud storage activity is done from the
background loop as previously.